sem10/DS-lab3/requirements.txt
PyYAML~=6.0
aio-pika~=9.0.4
